third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/origin-whitelisting-removal-expected.txt

CONSOLE ERROR: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/get.txt. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/get.txt. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/get.txt. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/get.txt. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/get.txt. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/get.txt. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/get.txt. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/get.txt. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
Tests the behavior of whitelisting origins and removing them later.

Testing: source origin: http://127.0.0.1:8000 destination origin: http:localhost 
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting NETWORK_ERR)
PASS: NetworkError: A network error occurred.
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting PASS)
PASS: PASS
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting NETWORK_ERR)
PASS: NetworkError: A network error occurred.


Testing: source origin: http://127.0.0.1:8000 destination origin: http:localhost allowing subdomains
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting NETWORK_ERR)
PASS: NetworkError: A network error occurred.
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting PASS)
PASS: PASS
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting NETWORK_ERR)
PASS: NetworkError: A network error occurred.


Testing: source origin: http://127.0.0.1:8000 destination origin: hTtP:LoCaLhOsT allowing subdomains
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting NETWORK_ERR)
PASS: NetworkError: A network error occurred.
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting PASS)
PASS: PASS
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting NETWORK_ERR)
PASS: NetworkError: A network error occurred.


Testing: source origin: http://127.0.0.1:8000 destination origin: http: allowing subdomains
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting NETWORK_ERR)
PASS: NetworkError: A network error occurred.
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting PASS)
PASS: PASS
Loading: http://localhost:8000/xmlhttprequest/resources/get.txt (expecting NETWORK_ERR)
PASS: NetworkError: A network error occurred.
